backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
sqlalchemy>=1.4.0
pydantic>=2.0.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv
requests
APScheduler>=3.10.0
asyncpg
